requirements.txt
# requirements.txt
# Python dependencies for c4c4 - Channel 4 Core Architecture DSL Generator

# LeanIX GraphQL integration
gql>=3.4.0
requests>=2.31.0
requests-toolbelt>=1.0.0
python-dotenv>=1.0.0

# Data validation and parsing
pydantic>=2.0.0